+++ ft_memset.c
@@ -6,18 +6,19 @@
 /*   By: gcosta-d <gcosta-d@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/20 16:35:38 by gcosta-d          #+#    #+#             */
-/*   Updated: 2021/08/28 00:00:53 by gcosta-d         ###   ########.fr       */
+/*   Updated: 2021/08/29 20:03:22 by gcosta-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void	*s, int	c, unsigned long	n)
+#include "libft.h"
+
+void	*ft_memset(void	*s, int	c, size_t	n)
 {
-	unsigned long	i;
+	size_t	i;
 	char	*buff;
 
 	i = 0;
 	buff = s;
-
 	while (i < n)
 	{
 		buff[i] = c;
